Fix MacOS

+++ engine/fonts.h
@@ -152,7 +152,7 @@ void main() {
                     texture,
                     glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                     glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                    face->glyph->advance.x
+                    (GLint) face->glyph->advance.x
             };
             return true;
         }
